site/requirements.txt
Flask==3.1.0
psycopg2-binary==2.9.10
elasticsearch==8.17.2
requests==2.32.3